moved log_message_transfer to logger module
